+++ src/connection.c
@@ -3546,6 +3546,51 @@ Connection_file_control(Connection *self, PyObject *const *fast_args, Py_ssize_t
   Py_RETURN_TRUE;
 }
 
+/** .. method:: vfsname(dbname: str) -> str | None
+
+Issues the `SQLITE_FCNTL_VFSNAME
+<https://sqlite.org/c3ref/c_fcntl_begin_atomic_write.html#sqlitefcntlvfsname>`__
+file control against the named database (`main`, `temp`, attached
+name).
+
+This is useful to see which VFS is in use, and if inheritance is used
+then ``/`` will separate the names.  If you have a :class:`VFSFile` in
+use then its fully qualified class name will also be included.
+
+If ``SQLITE_FCNTL_VFSNAME`` is not implemented, ``dbname`` is not a
+database name, or an error occurred then ``None`` is returned.
+*/
+static PyObject *
+Connection_vfsname(Connection *self, PyObject *const *fast_args, Py_ssize_t fast_nargs, PyObject *fast_kwnames)
+{
+
+  const char *dbname = NULL;
+
+  CHECK_USE(NULL);
+  CHECK_CLOSED(self, NULL);
+
+  {
+    Connection_vfsname_CHECK;
+    ARG_PROLOG(1, Connection_vfsname_KWNAMES);
+    ARG_MANDATORY ARG_str(dbname);
+    ARG_EPILOG(NULL, Connection_vfsname_USAGE, );
+  }
+
+  const char *vfsname = NULL;
+
+  /* because it is diagnostic and we can tell from vfsname changing and
+  because SQLite shell itself ignores the return code, we do the same */
+
+  PYSQLITE_VOID_CALL(sqlite3_file_control(self->db, dbname, SQLITE_FCNTL_VFSNAME, &vfsname));
+
+  PyObject *res = convertutf8string(vfsname);
+
+  if (vfsname)
+    sqlite3_free((void *)vfsname);
+
+  return res;
+}
+
 /** .. method:: sqlite3_pointer() -> int
 
 Returns the underlying `sqlite3 *
@@ -5245,6 +5290,8 @@ static PyMethodDef Connection_methods[] = {
      Connection_backup_DOC},
     {"file_control", (PyCFunction)Connection_file_control, METH_FASTCALL | METH_KEYWORDS,
      Connection_file_control_DOC},
+    {"vfsname", (PyCFunction)Connection_vfsname, METH_FASTCALL | METH_KEYWORDS,
+     Connection_vfsname_DOC},
     {"sqlite3_pointer", (PyCFunction)Connection_sqlite3_pointer, METH_NOARGS,
      Connection_sqlite3_pointer_DOC},
     {"set_exec_trace", (PyCFunction)Connection_set_exec_trace, METH_FASTCALL | METH_KEYWORDS,