+++ src/connection.c
@@ -4429,7 +4429,7 @@ fail:
 }
 
 static PyObject *formatsqlvalue(PyObject *Py_UNUSED(self), PyObject *value);
-/** .. method:: pragma(name: str, value: Optional[SQLiteValue] = None) -> Any
+/** .. method:: pragma(name: str, value: Optional[SQLiteValue] = None, *, schema: Optional[str] = None) -> Any
 
   Issues the pragma (with the value if supplied) and returns the result with
   :attr:`the least amount of structure <Cursor.get>`.  For example
@@ -4443,6 +4443,9 @@ static PyObject *formatsqlvalue(PyObject *Py_UNUSED(self), PyObject *value);
   may also contain encryption keys.  This method ensures they are
   not cached to avoid problems.
 
+  Use the `schema` parameter to run the pragma against a different
+  attached database (eg ``temp``).
+
   * :ref:`Example <example_pragma>`
 */
 static PyObject *
@@ -4450,6 +4453,7 @@ Connection_pragma(Connection *self, PyObject *const *fast_args, Py_ssize_t fast_
 {
   const char *name = NULL;
   PyObject *value = NULL;
+  const char *schema = NULL;
 
   CHECK_USE(NULL);
   CHECK_CLOSED(self, NULL);
@@ -4459,32 +4463,54 @@ Connection_pragma(Connection *self, PyObject *const *fast_args, Py_ssize_t fast_
     ARG_PROLOG(2, Connection_pragma_KWNAMES);
     ARG_MANDATORY ARG_str(name);
     ARG_OPTIONAL ARG_pyobject(value);
+    ARG_OPTIONAL ARG_optional_str(schema);
     ARG_EPILOG(NULL, Connection_pragma_USAGE, );
   }
 
-  PyObject *query = NULL, *value_str = NULL, *cursor = NULL, *res = NULL;
+  PyObject *value_format = NULL, *res = NULL, *cursor = NULL, *query_py = NULL;
+  const char *value_str = NULL;
+  char *query = NULL;
+
   if (value)
   {
-    value_str = formatsqlvalue(NULL, value);
+    value_format = formatsqlvalue(NULL, value);
+    if (!value_format)
+      goto error;
+    value_str = PyUnicode_AsUTF8(value_format);
     if (!value_str)
       goto error;
-    const char *utf8 = PyUnicode_AsUTF8(value_str);
-    if (!utf8)
-      goto error;
-
-    /* the form name(value) is used not name=value because some
-       pragmas like index_info work the former way while as do all that
-       support = */
-    query = PyUnicode_FromFormat("pragma %s(%s)", name, utf8);
   }
-  else
-    query = PyUnicode_FromFormat("pragma %s", name);
+
+  /* the form name(value) is used not name=value because some
+     pragmas like index_info only work that way, and all
+     support the parenthese method */
+
+  query = sqlite3_mprintf("pragma %s%w%s%s\"%w\"%s%s%s",
+                          /* surround schema with double quotes and follow with dot if set */
+                          schema ? "\"" : "",
+                          schema ? schema : "",
+                          schema ? "\"" : "",
+                          schema ? "." : "",
+                          /* pragma */
+                          name,
+                          /* value surrounded by parens if set */
+                          value_str ? "(" : "",
+                          value_str ? value_str : "",
+                          value_str ? ")" : "");
+
   if (!query)
+  {
+    PyErr_NoMemory();
+    goto error;
+  }
+
+  query_py = PyUnicode_FromString(query);
+  if (!query_py)
     goto error;
 
-  PyObject *vargs[] = {NULL, query, Py_False};
+  PyObject *vargs[] = {NULL, query_py, Py_False};
   PyObject *kwnames = PyTuple_Pack(1, apst.can_cache);
-  if(kwnames)
+  if (kwnames)
     cursor = Connection_execute(self, vargs + 1, 1 | PY_VECTORCALL_ARGUMENTS_OFFSET, kwnames);
   Py_XDECREF(kwnames);
   if (!cursor || !kwnames)
@@ -4493,10 +4519,12 @@ Connection_pragma(Connection *self, PyObject *const *fast_args, Py_ssize_t fast_
   res = PyObject_GetAttr(cursor, apst.get);
 
 error:
-  Py_XDECREF(query);
-  Py_XDECREF(value_str);
+  Py_XDECREF(value_format);
   Py_XDECREF(cursor);
+  Py_XDECREF(query_py);
+  sqlite3_free(query);
 
+  assert((res && !PyErr_Occurred()) || (!res && PyErr_Occurred()));
   return res;
 }