+++ doc/changes.rst
@@ -10,6 +10,13 @@ history <https://devguide.python.org/versions/>`__.
 APSW changes by version
 -----------------------
 
+next
+====
+
+Address how errors are handled in VFS xRandomness routine, that is
+only called once by SQLite to seed its random number generator.
+(:issue:`526`)
+
 3.46.0.0
 ========