+++ doc/changes.rst
@@ -13,6 +13,9 @@ APSW changes by version
 next
 ====
 
+:func:`apsw.ext.query_info` provides the count and names of bindings
+parameters.  (:issue:`528`)
+
 Address how errors are handled in VFS xRandomness routine, that is
 only called once by SQLite to seed its random number generator.
 (:issue:`526`)